Use KnownTypesBindire with Object
